Move utils folder to project root.
