+++ src/js/App/ErrorComponents/GatewayErrorComponent.tsx
@@ -8,7 +8,7 @@ import ChromeLink from '../Sidenav/Navigation/ChromeLink';
 import { Text, TextContent } from '@patternfly/react-core';
 import { useIntl } from 'react-intl';
 import Messages from '../../Messages';
-import { ThreeScaleError } from '../../utils/responseInterceptors';
+import { ThreeScaleError } from '../../../utils/responseInterceptors';
 
 export type GatewayErrorComponentProps = {
   error: ThreeScaleError;
