+++ src/js/App/Sidenav/Navigation/ChromeNavItem.js
@@ -11,7 +11,7 @@ import { isBeta } from '../../../utils';
 import { betaBadge } from '../../Header/Tools';
 import ChromeLink from './ChromeLink';
 import { useDispatch, useSelector } from 'react-redux';
-import useRenderFedramp from '../../../utils/useRenderFedramp';
+import useRenderFedramp from '../../../../utils/useRenderFedramp';
 import { markActiveProduct } from '../../../redux/actions';
 
 const ChromeNavItem = ({ appId, className, href, isHidden, ignoreCase, title, isExternal, isBeta: isBetaEnv, active, product, notifier = '' }) => {
