+++ src/js/redux/reducers.ts
@@ -3,7 +3,7 @@ import { ChromeUser } from '@redhat-cloud-services/types';
 import { REQUESTS_COUNT, REQUESTS_DATA } from '../consts';
 import { NavItem } from '../types';
 import { generateRoutesList, highlightItems, isBeta, isFedRamp, levelArray } from '../utils';
-import { ThreeScaleError } from '../utils/responseInterceptors';
+import { ThreeScaleError } from '../../utils/responseInterceptors';
 import { AccessRequest, ChromeModule, ChromeState, Navigation } from './store';
 
 export function contextSwitcherBannerReducer(state: ChromeState): ChromeState {
