+++ src/utils/isAnsibleTrialFlagActive.ts
@@ -1,4 +1,4 @@
-import logger from '../jwt/logger';
+import logger from '../js/jwt/logger';
 
 export const ANSIBLE_TRIAL_FLAG = 'chrome.ansible.trial';
 const TRIAL_DURATION = 10 * 60 * 1000; // 10 minutes
