+++ src/utils/useNavigation.js
@@ -2,8 +2,8 @@ import axios from 'axios';
 import { useContext, useEffect, useRef } from 'react';
 import { batch, useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { loadLeftNavSegment, setGatewayError } from '../redux/actions';
-import { isBeta } from '../utils';
+import { loadLeftNavSegment, setGatewayError } from '../js/redux/actions';
+import { isBeta } from '../js/utils';
 import { evaluateVisibility } from './isNavItemVisible';
 import { QuickStartContext } from '@patternfly/quickstarts';
 import { useFlagsStatus } from '@unleash/proxy-client-react';
